Update AgentProfile with popover & menu
