+++ modules/hitl2/src/views/full/Components/AgentProfile.tsx
@@ -1,51 +1,45 @@
-import React, { FC, useContext } from 'react'
+import React, { FC } from 'react'
 
-import { Context } from './../app/Store'
-
-import { ApiType } from './../Api'
 import { AgentType } from './../../../types'
 
-import { Button } from '@blueprintjs/core'
-import { toast, lang } from 'botpress/shared'
+import { Popover, Position, Menu, MenuItem, Button, Icon } from '@blueprintjs/core'
+import { lang } from 'botpress/shared'
+
+import AgentIcon from './AgentIcon'
 
 type Props = {
-  api: ApiType
+  toggleOnline: (online) => {}
+  loading: boolean
 } & Partial<AgentType>
 
 const AgentProfile: FC<Props> = props => {
-  const { dispatch } = useContext(Context)
-
-  async function toggleOnline(online: boolean) {
-    try {
-      let agent
-      if (online) {
-        agent = await props.api.setOnline()
-      } else {
-        agent = await props.api.setOffline()
-      }
-      dispatch({ type: 'setCurrentAgent', payload: agent }) // optimistic update, will also be updated via websocket event
-      online
-        ? toast.success(lang.tr('module.hitl2.agent.onlineSuccess'))
-        : toast.success(lang.tr('module.hitl2.agent.offlineSuccess'))
-    } catch (error) {
-      dispatch({ type: 'setError', payload: error })
-    }
-  }
-
   return (
-    <div>
-      <ul>
-        <li>Online: {props.online ? 'online' : 'offline'}</li>
-        <li>Id: {props.id}</li>
-        <li>Name: {props.fullName}</li>
-      </ul>
-
-      {props.online ? (
-        <Button onClick={() => toggleOnline(false)}>{lang.tr('module.hitl2.agent.getOffline')}</Button>
-      ) : (
-        <Button onClick={() => toggleOnline(true)}>{lang.tr('module.hitl2.agent.getOnline')}</Button>
-      )}
-    </div>
+    <Popover
+      content={
+        <Menu>
+          {props.online ? (
+            <MenuItem
+              onClick={() => props.toggleOnline(false)}
+              text={lang.tr('module.hitl2.agent.getOffline')}
+            ></MenuItem>
+          ) : (
+            <MenuItem
+              onClick={() => props.toggleOnline(true)}
+              text={lang.tr('module.hitl2.agent.getOnline')}
+            ></MenuItem>
+          )}
+        </Menu>
+      }
+      position={Position.BOTTOM_RIGHT}
+    >
+      <Button loading={props.loading} minimal={true} style={{ position: 'relative' }}>
+        <AgentIcon online={props.online} />
+        <span style={{ paddingLeft: 10 }}>
+          {props?.online ? lang.tr('module.hitl2.agent.online') : lang.tr('module.hitl2.agent.offline')}
+        </span>
+        <Icon icon="caret-down"></Icon>
+      </Button>
+    </Popover>
   )
 }
